nvim: Move floating term keymap to buffer
Previously, the floating terminal keymap was attached to the global
namespace. This overwrote other keymaps and also applied it to all
terminals. This made <Esc> open and close the floating terminal when in
a terminal, which is not intended.

The keybinding is now attached to the buffer and created at
instantiation. This may be an issue later if I refactor and move the
keybindings out to the standard .vimrc.